Fix the HIP kernel compile, and put HIP back in the default device set
The kernels failed on the first architecture with "non-const lvalue reference
cannot bind to a temporary", pointing into AMD's amd_hip_vector_types.h. It
looked like a ROCm bug, it reproduced on 6.4 and 7.1, and HIP was taken out of
the default set because one error takes out all 18 architectures - they share a
single custom build rule.

It was ours, in one line of svm_rhino_procedurals.h:

    color /= (colors_added == 0 ? 1 : colors_added);

color is a float4 and the divisor an int. GPU builds use native vector types, so
on HIP float4 is AMD's HIP_vector_type<float, 4>. An int divisor deduces HIP's
member template operator/=(U) exactly and so beats the float4 overloads in
util/math_float4.h, which would need int -> float. That template returns
HIP_vector_type& and its body is

    return *this /= HIP_vector_type{x};

and 5.2 added operator/=(float4 &a, const float4 b), a better match for that
inner expression than HIP's own member - which returns float4 by value.
Returning that temporary as a reference is the error. 3.5 declared only the
float overload, so the inner division stayed on HIP's member and compiled.

Upstream never trips it: nothing in Cycles divides a float4 by an int in device
code. Casting the divisor to float is identical arithmetic everywhere and just
stops the overload hijack.

All 18 architectures now build against ROCm 6.4, gfx1150 included, so HIP joins
the automatic set again. Compiling is not rendering - the HIP compiler has a
history of miscompiling Cycles kernels on particular devices, which is what the
-O1 workaround in the kernel CMakeLists is for, so this still wants a render
test on real AMD hardware.

+++ src/kernel/svm/svm_rhino_procedurals.h
@@ -3666,7 +3666,27 @@ ccl_device float4 dots_texture(KernelGlobals kg,
   }
 
   if (composition_type == RHINO_DOTS_COMPOSITION_AVERAGE) {
-    color /= (colors_added == 0 ? 1 : colors_added);
+    /* Divide by a float, not an int. This one line cost HIP every kernel it has.
+     *
+     * On HIP, float4 is AMD's own HIP_vector_type<float, 4> - GPU builds use the
+     * native vector types. An int divisor deduces HIP's member template
+     * operator/=(U) exactly, so it beats the float4 overloads in
+     * util/math_float4.h, which would need an int -> float conversion. That
+     * template returns HIP_vector_type& and its body is
+     *
+     *     return *this /= HIP_vector_type{x};
+     *
+     * and 5.2 added operator/=(float4 &a, const float4 b), which is a better
+     * match for that inner expression than HIP's own member - and it returns
+     * float4 by value. Returning that temporary from a function declared to
+     * return a non-const reference is the error, and it surfaces inside AMD's
+     * header with nothing pointing back at this file. Cycles 3.5 declared only
+     * the float overload, so the inner division stayed on HIP's member.
+     *
+     * Upstream never trips it: nothing in Cycles divides a float4 by an int in
+     * device code. Only this procedural does. All 18 architectures build from
+     * one custom-build rule, so this single error meant no HIP kernels at all. */
+    color /= (float)(colors_added == 0 ? 1 : colors_added);
   }
   else if (composition_type == RHINO_DOTS_COMPOSITION_STANDARD) {
     color = (1.f - standard_composition_value) * original_background_color +
